backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.9.2
websockets==13.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
ffmpeg-python==0.2.0
speechbrain>=1.0.0
scikit-learn==1.3.2
librosa==0.10.1
llama-cpp-python==0.2.20
openai==1.3.0
transformers==4.35.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
requests==2.31.0


